refactor: 코드 일부 수정

+++ bj_1978/bj_1978_lyj.java
@@ -18,7 +18,7 @@ public class bj_1978_lyj {
         if (n == 2)
             return true;
 
-        for (int i = 2; i < (Math.sqrt(n) + 1); i++) {
+        for (int i = 2; i < (int)Math.sqrt(n) + 1; i++) {
             if (n % i == 0)
                 return false;
         }
